fix compilation warning and enforce that both slit motors be defined (could consider allowing only one, with limited functionality, but  is that necessary?)
dh

	modified:   slitd/slit_interface.cpp
	modified:   slitd/slit_interface.h

+++ slitd/slit_interface.h
@@ -106,7 +106,7 @@ namespace Slit {
   class Interface {
     private:
       bool class_initialized;
-      int numdev;
+      size_t numdev;
       float maxwidth;
       float minwidth;
     public:
